finishing transaction

+++ api/dto/product.go
@@ -1,27 +1,27 @@
 package dto
 
 type CreateProductRequest struct {
-	Name        string `form:"name" validate:"required"`
-	Stock       int    `form:"stock" validate:"required"`
-	Price       int    `form:"price" validate:"required"`
-	Description string `form:"description" validate:"required"`
-	Image       string `form:"image" validate:"required"`
+	Name        string  `form:"name" validate:"required"`
+	Stock       int     `form:"stock" validate:"required"`
+	Price       float64 `form:"price" validate:"required"`
+	Description string  `form:"description" validate:"required"`
+	Image       string  `form:"image" validate:"required"`
 }
 
 type UpdateProductRequest struct {
-	Name        string `form:"name"`
-	Stock       int    `form:"stock"`
-	Price       int    `form:"price"`
-	Description string `form:"description"`
-	Image       string `form:"image"`
+	Name        string  `form:"name"`
+	Stock       int     `form:"stock"`
+	Price       float64 `form:"price"`
+	Description string  `form:"description"`
+	Image       string  `form:"image"`
 }
 
 type ProductResponse struct {
-	ID          int    `json:"id"`
-	Name        string `json:"name"`
-	Price       int    `json:"price"`
-	Stock       int    `json:"stock"`
-	Image       string `json:"image"`
-	Description string `json:"description"`
-	Status      string `json:"status"`
+	ID          int     `json:"id"`
+	Name        string  `json:"name"`
+	Price       float64 `json:"price"`
+	Stock       int     `json:"stock"`
+	Image       string  `json:"image"`
+	Description string  `json:"description"`
+	Status      string  `json:"status"`
 }
