+++ api/dto/transaction.go
@@ -9,21 +9,21 @@ type ProductRequestForTransaction struct {
 }
 
 type ProductResponseForTransaction struct {
-	ID          int    `json:"id"`
-	Name        string `json:"name"`
-	Price       int    `json:"price"`
-	Image       string `json:"image"`
-	Description string `json:"description"`
-	OrderQty    int    `json:"orderQty"`
+	ID          int     `json:"id"`
+	Name        string  `json:"name"`
+	Price       float64 `json:"price"`
+	Image       string  `json:"image"`
+	Description string  `json:"description"`
+	OrderQty    int     `json:"orderQty"`
 }
 
 type CreateTransactionRequest struct {
-	Total    int                            `json:"total" validate:"required"`
+	Total    float64                        `json:"total" validate:"required"`
 	UserID   int                            `json:"user_id" validate:"required"`
 	Products []ProductRequestForTransaction `json:"products" validate:"required"`
 }
 type UpdateTransactionRequest struct {
-	Total    int                            `json:"total"`
+	Total    float64                        `json:"total"`
 	UserID   int                            `json:"user_id"`
 	Products []ProductRequestForTransaction `json:"products"`
 	Status   string                         `json:"status"`
@@ -33,7 +33,7 @@ type TransactionResponse struct {
 	ID         string                          `json:"id"`
 	MidtransID string                          `json:"midtrans_id"`
 	OrderDate  string                          `json:"order_date"`
-	Total      int                             `json:"total"`
+	Total      float64                         `json:"total"`
 	Status     string                          `json:"status"`
 	User       models.UserResponse             `json:"user"`
 	Products   []ProductResponseForTransaction `json:"products"`
